ingestion/requirements.txt
snowflake-connector-python==3.10.1
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.4
